# Caso 1 - Parte  2/src/backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DUA Streamliner API
  version: 0.1.0
  description: >
    REST contract (MVP). Source of truth for public API; keep in sync with FastAPI routers.
servers:
  - url: https://api.example.com
paths:
  /health/liveness:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: OK
  /health/readiness:
    get:
      summary: Readiness probe
      responses:
        "200":
          description: OK
  /v1/jobs:
    post:
      summary: Create job and obtain upload URLs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [files, template_id]
              properties:
                template_id:
                  type: string
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      size: { type: integer }
      responses:
        "201":
          description: Created
  /v1/jobs/{job_id}/submit:
    post:
      summary: Submit job after uploads complete
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Accepted
  /v1/jobs/{job_id}:
    get:
      summary: Get job status
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
  /v1/jobs/{job_id}/progress:
    get:
      summary: SSE progress stream
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Event stream
  /v1/jobs/{job_id}/download:
    get:
      summary: Pre-signed download URL for DUA Word output
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT